Add services, controllers and views
Add services, controllers and views for entities, fix small bugs,
add queries for getting number of records

+++ src/main/java/dev/andrylat/carsharing/controllers/BodyTypesController.java
@@ -32,7 +32,7 @@ public class BodyTypesController {
     }
 
     @GetMapping("/{id}")
-    public String getById(@PathVariable("id") int id, Model model) {
+    public String getById(@PathVariable("id") long id, Model model) {
         model.addAttribute("bodyType", bodyTypeService.getById(id));
         return "bodytypes/getById";
     }
